Remove SQLite from Shared tables test
# Conflicts:
#	composer.lock

+++ tests/e2e/Adapter/Base.php
@@ -1396,10 +1396,14 @@ abstract class Base extends TestCase
     {
         $documents = static::getDatabase()->find('documents');
         $document = $documents[0];
+        $this->assertArrayHasKey('$id', $document);
         $this->assertArrayNotHasKey('$tenant', $document);
 
         $document = static::getDatabase()->getDocument('documents', $document->getId());
+        $this->assertArrayHasKey('$id', $document);
         $this->assertArrayNotHasKey('$tenant', $document);
+
+        $this->assertEquals(true, false);
     }
 
     public function testEmptySearch(): void
@@ -12625,6 +12629,11 @@ abstract class Base extends TestCase
          */
         $database = static::getDatabase();
 
+        if (!$database->getAdapter()->getSupportForSchemas()) {
+            $this->expectNotToPerformAssertions();
+            return;
+        }
+
         if ($database->exists('schema1')) {
             $database->setDatabase('schema1')->delete();
         }
@@ -12643,18 +12652,14 @@ abstract class Base extends TestCase
             ->setNamespace('')
             ->create();
 
-        if ($database->getAdapter()->getSupportForSchemas()) {
-            $this->assertEquals(true, $database->exists('schema1'));
-        }
+        $this->assertEquals(true, $database->exists('schema1'));
 
         $database
             ->setDatabase('schema2')
             ->setNamespace('')
             ->create();
 
-        if ($database->getAdapter()->getSupportForSchemas()) {
-            $this->assertEquals(true, $database->exists('schema2'));
-        }
+        $this->assertEquals(true, $database->exists('schema2'));
 
         /**
          * Table
@@ -12670,9 +12675,7 @@ abstract class Base extends TestCase
             ->setTenant($tenant1)
             ->create();
 
-        if ($database->getAdapter()->getSupportForSchemas()) {
-            $this->assertEquals(true, $database->exists('sharedTables'));
-        }
+        $this->assertEquals(true, $database->exists('sharedTables'));
 
         $database->createCollection('people', [
             new Document([
